+++ src/main/native/include/ctre/phoenix/motorcontrol/can/BaseMotorController.h
@@ -106,8 +106,8 @@ struct SlotConfiguration{
 		kI(0.0),
 		kD(0.0),
 		kF(0.0),
-		integralZone(0.0), 
-		allowableClosedloopError(0.0), 
+		integralZone(0), 
+		allowableClosedloopError(0), 
 		maxIntegralAccumulator(0.0),
 		closedLoopPeakOutput(1.0),
 		closedLoopPeriod(1)
@@ -388,10 +388,10 @@ public:
 	virtual ctre::phoenix::ErrorCode SetSelectedSensorPosition(int sensorPos, int pidIdx = 0, int timeoutMs = 50);
 	//------ status frame period changes ----------//
 	virtual ctre::phoenix::ErrorCode SetControlFramePeriod(ControlFrame frame, int periodMs);
-	virtual ctre::phoenix::ErrorCode SetStatusFramePeriod(StatusFrame frame, int periodMs,
+	virtual ctre::phoenix::ErrorCode SetStatusFramePeriod(StatusFrame frame, uint8_t periodMs,
 			int timeoutMs = 0);
 	virtual ctre::phoenix::ErrorCode SetStatusFramePeriod(StatusFrameEnhanced frame,
-			int periodMs, int timeoutMs = 0);
+			uint8_t periodMs, int timeoutMs = 0);
 	virtual int GetStatusFramePeriod(StatusFrame frame, int timeoutMs = 0);
 	virtual int GetStatusFramePeriod(StatusFrameEnhanced frame, int timeoutMs = 0);
 	//----- velocity signal conditionaing ------//
